+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 <a name="0.9.18"><a/>
 # <angular/> 0.9.18 jiggling-armfat (in-progress) #
 
+### Bug Fixes
+- Issue #449: [ng:options] should support binding to a property of an item.
+- Issue #464: [ng:options] incorrectly re-grew options on datasource change
+- Issue #448: [ng:options] should support iterating over objects
+- Issue #463: [ng:options] should support firing ng:change event
+- Issue #450: [ng:options] should support group by (select option groups)